Add generic port configuration and refactor
