+++ include/slac/channel.hpp
@@ -3,9 +3,7 @@
 #ifndef SLAC_CHANNEL_HPP
 #define SLAC_CHANNEL_HPP
 
-#ifdef ESP_PLATFORM
-#include "port/esp32s3/port_config.hpp"
-#endif
+#include "port/generic/port_config.hpp"
 
 #include <slac/transport.hpp>
 #include <string>
