+++ include/slac/slac.hpp
@@ -3,9 +3,7 @@
 #ifndef SLAC_SLAC_HPP
 #define SLAC_SLAC_HPP
 
-#ifdef ESP_PLATFORM
-#include "port/esp32s3/port_config.hpp"
-#endif
+#include "port/generic/port_config.hpp"
 
 #include <cstdint>
 #include <utility>
