+++ include/slac/transport.hpp
@@ -1,9 +1,7 @@
 #ifndef SLAC_TRANSPORT_HPP
 #define SLAC_TRANSPORT_HPP
 
-#ifdef ESP_PLATFORM
-#include "port/esp32s3/port_config.hpp"
-#endif
+#include "port/generic/port_config.hpp"
 
 #include <cstdint>
 #include <cstddef>
