+++ port/esp32s3/qca7000.cpp
@@ -1,4 +1,5 @@
 #include "qca7000.hpp"
+#include "../generic/port_config.hpp"
 #include "port_config.hpp"
 #ifdef ESP_PLATFORM
 #include <esp_log.h>
@@ -8,6 +9,7 @@
 #include <arpa/inet.h>
 #define ESP_LOGE(tag, fmt, ...)
 #define ESP_LOGI(tag, fmt, ...)
+#define ESP_LOGW(tag, fmt, ...)
 static inline uint32_t esp_random() {
     return 0x12345678u;
 }
