+++ port/esp32s3/qca7000_link.hpp
@@ -1,6 +1,7 @@
 #ifndef SLAC_QCA7000_LINK_HPP
 #define SLAC_QCA7000_LINK_HPP
 
+#include "../generic/port_config.hpp"
 #ifdef ESP_PLATFORM
 #include "port_config.hpp"
 #endif
